Conflicts:

	htdocs/admin/index.php
	htdocs/lib/db/install.xml
	htdocs/lib/db/upgrade.php
	htdocs/lib/version.php
	htdocs/lib/view.php

+++ htdocs/artefact/blog/import/leap/lib.php
@@ -28,9 +28,9 @@
 defined('INTERNAL') || die();
 
 /**
- * Implements LEAP2A import of blog related entries into Mahara
+ * Implements Leap2A import of blog related entries into Mahara
  *
- * For more information about LEAP blog importing, see:
+ * For more information about Leap blog importing, see:
  * http://wiki.mahara.org/Developer_Area/Import//Export/LEAP_Import/Blog_Artefact_Plugin
  *
  * TODO:
@@ -103,10 +103,17 @@ class LeapImportBlog extends LeapImportArtefactPlugin {
         }
         else {
             // The blog can import any entry as a literal blog post
+            // Get files that this blogpost/catchall feels are a part of it
+            $otherrequiredentries = array();
+            foreach ($entry->link as $link) {
+                if ($importer->curie_equals($link['rel'], '', 'enclosure') && isset($link['href'])) {
+                    $otherrequiredentries[] = (string)$link['href'];
+                }
+            }
             $strategies[] = array(
                 'strategy' => self::STRATEGY_IMPORT_AS_ENTRY,
                 'score'    => 10,
-                'other_required_entries' => array(),
+                'other_required_entries' => $otherrequiredentries,
             );
         }
 
@@ -185,7 +192,7 @@ class LeapImportBlog extends LeapImportArtefactPlugin {
                         $artefactids = $importer->get_artefactids_imported_by_entryid((string)$blogpostlink['href']);
                         if (isset($artefactids[0])) {
                             $blogpost->attach($artefactids[0]);
-                        } else { // it may be just an attached file, with no leap2a element in its own right ....
+                        } else { // it may be just an attached file, with no Leap2A element in its own right ....
                             if ($id = self::attach_linked_file($blogpostentry, $blogpostlink, $importer)) {
                                 $blogpost->attach($id);
                                 $newartefactmapping[(string)$blogpostlink['href']][] = $id;
@@ -201,6 +208,25 @@ class LeapImportBlog extends LeapImportArtefactPlugin {
             }
             break;
         case self::STRATEGY_IMPORT_AS_ENTRY:
+            $blogpostids = $importer->get_artefactids_imported_by_entryid((string)$entry->id);
+            if (!isset($blogpostids[0])) {
+                // weird!
+                break;
+            }
+            $blogpost = new ArtefactTypeBlogPost($blogpostids[0]);
+            foreach ($entry->link as $link) {
+                if ($importer->curie_equals($link['rel'], '', 'enclosure') && isset($link['href'])) {
+                    if (isset($artefactids[0])) {
+                        $blogpost->attach($artefactids[0]);
+                    } else {
+                        if ($id = self::attach_linked_file($entry, $link, $importer)) {
+                            $blogpost->attach($id);
+                            $newartefactmapping[(string)$link['href']][] = $id;
+                        }
+                    }
+                }
+            }
+            $blogpost->commit();
             self::setup_outoflinecontent_relationship($entry, $importer);
             break;
         default:
@@ -210,8 +236,8 @@ class LeapImportBlog extends LeapImportArtefactPlugin {
     }
 
     /**
-     * Attaches a file to a blogpost entry that was just linked directly, rather than having a leap2a entry
-     * See http://wiki.cetis.ac.uk/LEAP2A_relationships#Attachments
+     * Attaches a file to a blogpost entry that was just linked directly, rather than having a Leap2a entry
+     * See http://wiki.leapspecs.org/2A/files
      *
      * @param SimpleXMLElement $blogpostentry
      * @param SimpleXMLElement $blogpostlink
@@ -241,7 +267,9 @@ class LeapImportBlog extends LeapImportArtefactPlugin {
     public static function cleanup(PluginImportLeap $importer) {
         if (self::$importedablog && self::$firstblogid) {
             $blog = artefact_instance_from_id(self::$firstblogid);
-            $blog->delete();
+            if (!$blog->has_children()) { // TODO see #544160
+                $blog->delete();
+            }
         }
     }
